requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
shapely
pydantic
python-dotenv
scikit-learn
pandas
numpy
joblib
google-generativeai